move cli code for unified_diff
